+++ src/test/java/com/global/api/tests/terminals/upa/UpaGiftTests.java
@@ -4,6 +4,7 @@ import com.global.api.entities.enums.ConnectionModes;
 import com.global.api.entities.enums.DeviceType;
 import com.global.api.entities.enums.TransactionType;
 import com.global.api.entities.exceptions.ApiException;
+import com.global.api.logging.RequestFileLogger;
 import com.global.api.services.DeviceService;
 import com.global.api.terminals.ConnectionConfig;
 import com.global.api.terminals.TerminalResponse;
@@ -11,7 +12,6 @@ import com.global.api.terminals.abstractions.IDeviceInterface;
 import com.global.api.terminals.abstractions.IDeviceResponse;
 import com.global.api.terminals.upa.Entities.Enums.UpaCardTypeFilter;
 import com.global.api.tests.terminals.hpa.RandomIdProvider;
-import com.global.api.logging.RequestFileLogger;
 import org.junit.FixMethodOrder;
 import org.junit.Test;
 import org.junit.runners.MethodSorters;
@@ -28,7 +28,7 @@ public class UpaGiftTests {
     public UpaGiftTests() throws ApiException {
         ConnectionConfig config = new ConnectionConfig();
         config.setPort(8080);
-        config.setIpAddress("10.253.146.155");
+        config.setIpAddress("10.27.188.113");
         config.setTimeout(15000);
         config.setRequestIdProvider(new RandomIdProvider());
         config.setDeviceType(DeviceType.UPA_DEVICE);
@@ -45,9 +45,8 @@ public class UpaGiftTests {
     @Test
     public void giftAddValue() throws ApiException {
         TerminalResponse response = device.giftAddValue(new BigDecimal("10.01"))
-            .withGiftTransactionType(TransactionType.Sale)
-            .execute();
-
+                .withGiftTransactionType(TransactionType.Sale)
+                .execute();
         runBasicTests(response);
         assertEquals("5022440000000000007", response.getUnmaskedCardNumber());
     }